Made lots of changes in models and APIs

+++ controllers/taskController.js
@@ -1,45 +1,53 @@
 const AppError = require('../utils/AppError');
 const Task = require('./../models/taskModel');
 const catchAsync = require('./../utils/catchAsync');
+const User = require('./../models/userModel');
 
 exports.createTask = catchAsync(async (req,res,next)=>{
     const {date,task,status} = req.body;
+    const user = req.user;
         const newTask = await Task.create({
             date,
             task,
-            status
+            status,
+            user:user._id
         })
-    res.status(200).json({
-        status: 'Success',
-        data: {
-            newTask
+        if(!user.tasks){
+            user.tasks = []
         }
-    })
+        user.tasks.push(newTask);
+        await user.save({validateBeforeSave:false});
+        res.status(200).json({
+            newTask,
+            user
+        })
+    
 });
 
 exports.updateTask = catchAsync(async (req,res,next) => {
-        const selectedTask = await Task.findByIdAndUpdate(req.params.id,req.body,{
-            new: true
-        });
-        if(!selectedTask){
-            return next(new AppError('Task not found in the database!!',404))
+    if(!(req.user.tasks.includes(req.params.id))) {
+        return next(new AppError('Task not found for the user',404));
+    }
+    const selectedTask = await Task.findByIdAndUpdate(req.params.id,req.body,{
+        new:true
+    });
+    res.status(200).json({
+        status:'Success',
+        message:'Task updated successfully!!',
+        data:{
+            selectedTask
         }
-
-        res.status(200).json({
-            status:'Success',
-            message:'Task updated successfully!!',
-            data:{
-                selectedTask
-            }
-        })  
+    })          
     }
 )
 
 exports.deleteTask = catchAsync(async (req,res,next) =>{
-        const selectedTask = await Task.findByIdAndDelete(req.params.id);
-        if(!selectedTask){
-            return next(new AppError('Task not found in the database!!',404))
+        if(!(req.user.tasks.includes(req.params.id))) {
+            return next(new AppError('Task not found for the user',404));
         }
+        const selectedTask = await Task.findByIdAndDelete(req.params.id);
+        req.user.tasks.pop(req.params.id);
+        req.user.save({validateBeforeSave:false});
         res.status(200).json({
             status:'Success',
             message:'Task deleted successfully!',
@@ -49,15 +57,20 @@ exports.deleteTask = catchAsync(async (req,res,next) =>{
 )
 
 exports.getAllTasks = catchAsync(async (req,res,next) => {
-        const tasks = await Task.find();
-        if(!tasks){
-            return next(new AppError(' No Task found in the database!!',404))
+        let query = Task.find().where('user').equals(req.user._id);
+        const page = req.query.page * 1 || 1;
+        const limit = req.query.limit * 1 || 3;
+        const skip = (page - 1) * limit;
+        if(req.query.sort){
+            const sortBy = req.query.sort.split(',').join(" ");
+            query = query.sort(sortBy);
         }
+        query = query.skip(skip).limit(limit);
+        // const user = await User.findOne({_id : req.user._id}).populate({path:'tasks',select:'-__v -user'});
+        const tasks = await query;
         res.status(200).json({
             status:'Success',
-            data:{
-                tasks
-            }
+            tasks:tasks  
         })}
 )
